Improve window focus script
Updated the window focus script to better allow switching between
windows in a stack, or switching back and forth between the east and
west windows.

Example:
Space 1 has four windows, 2 stacked to the right, and 2 stacked to the
left. lets say you want to bring up the hidden window in the right
stack, moving the focus to the right (mod - d) to target the currently
top-of-stack window, and now simply switch to the next window in that
stack by pressing (mod - d) again. This brings the new window to the top
of the stack.
Now, switching back to the left pane is just a (mod - a) away, instead
of looping through all possible windows, which would cause the
top-of-stack window to change.